Fix spacing between link text and span (#3)
* fix(messages): don't display if messages are blank

We have certain icon decorations that do not need message decoration,
such as a chevron.

* fix(spacing): add encoded space for separation

When the span is created, the space that once was there is no longer,
and this can cause the last word in the span to appear as if it's right
next to the previous text.  This insert an encoded space in the link
that is not read by screen readers, yet gives the space needed for the
last word to look as if it's still attached to the link text as before.